[RPW] Fix default volume profile updating for file-locking version control systems
Fix for https://jira.unity3d.com/browse/UUM-33946.

When user opens HDRP Global Settings, the assigned  Default Volume Profile is updated to contain all VolumeComponents found in the project. Before this PR, the code contained a check to skip this update when the asset is not writable to handle the situation when the profile asset is inside a read-only package. This implementation missed the fact that the asset can be read-only also because the user project is using a locking version control system such as Perforce or Plastic. The check has been changed to allow updating also in this situation - the volume profile should now update and be automatically checked out in p4/plastic.