Fix SRP shader warnings related to half 1
Fix warnings in SRP shaders related with "conversion from larger type to smaller type". There are too many warnings to fix in one PR, so doing smaller PRs that can be easily tested and reviewed to avoid breaking anything. These warnings appear on Metal when defining half as minfloat and enabling Unified mode after the following changes:

```
--- a/Packages/com.unity.render-pipelines.core/ShaderLibrary/Common.hlsl
+++ b/Packages/com.unity.render-pipelines.core/ShaderLibrary/Common.hlsl
@@ -118,7 +118,7 @@
 // The including shader should define whether half
 // precision is suitable for its needs.  The shader
 // API (for now) can indicate whether half is possible.
-#if defined(SHADER_API_MOBILE) || defined(SHADER_API_SWITCH)
+#if defined(SHADER_API_MOBILE) || defined(SHADER_API_SWITCH) || defined(UNITY_UNIFIED_SHADER_PRECISION_MODEL)
 #define HAS_HALF 1
 #else
 #define HAS_HALF 0
@@ -134,7 +134,7 @@
 #define REAL_IS_HALF 0
 #endif // Do we have half?
 
-#if REAL_IS_HALF || (defined(UNITY_UNIFIED_SHADER_PRECISION_MODEL) && (defined(UNITY_COMPILER_HLSL) || defined(UNITY_COMPILER_DXC)))
+#if REAL_IS_HALF
 #define half min16float
 #define half2 min16float2
 #define half3 min16float3
 ```

The above changes are not included in this PR because they would generate too many warnings if Shader Precision Model is set to Unified.