[PS5] Make PlayStation5NGGC the default Graphics API
The Unity platform implementation for PS5 has 2 graphics API: PlayStation5 and PlayStation5NGGC.
Since in an effort to improve performance and modernize our graphics backend, we've landed the PlayStation5NGGC graphics API in 2021.3 and later.
We consider the NGGC graphics API production-ready, as proven by our PSVR2 customers, since it's the only API we support for the VR device.
We've consistently seen better performance on both CPU and GPU, and better memory usage, while providing the same visual quality and more features.
For these reasons, we believe PlayStation5NGGC is the better choice and should be the default.