[RPW] Port URP Global Settings to UITK, new DefaultVolumeProfileEditor
JIRA: https://jira.unity3d.com/browse/XPIPELINE-703

This PR implements URP Global Settings, and more importantly the DefaultVolumeProfileEditor using UITK. The main reason to do this (in addition to overall need to move to UITK) is to improve the performance of the SRP Global Settings window. Since there can be a very long list of VolumeComponentEditors, we want to take advantage of the virtualization feature of UITK, meaning that only the elements visible on screen are updated every frame. With IMGUI this was not possible, so large amount of VolumeComponentEditors could really tank the performance even though most of them were not visible. With this PR, the user experience when scrolling should be much smoother.

In this PR, only URP Global Settings uses the new UITK-based DefaultVolumeProfileEditor. HDRP still uses the old IMGUI implementation, so there are no performance improvements there.

The intent of this PR is that there are no visual or behavioral changes. Everything inside Project Settings > URP Global Settings should look like and work exactly as before, only with better performance.

Before (IMGUI):
![image](https://media.github.cds.internal.unity3d.com/user/3380/files/b915b167-7769-4a77-9211-d93e75255296)

After (UITK):
![image](https://media.github.cds.internal.unity3d.com/user/3380/files/9f1a0ff3-6dec-4143-b0a2-2fbb055ff4ff)