[2023.1] [URP] Strip whole shader exclusive to HDR output and cleaned keyword enabling
The `BlitHDROverlay.shader` on URP is unnecessary when HDR output is disallowed in the player settings, but was still included in the builds. The KDR output-related keywords were stripped but there were 24 variants still remaining. This PR strips all variants of the shader when HDR output is disallowed, meaning less variants are compiled.

Additionally the HDR output keyword management in the HDROutputUtils API was enabling keywords but relying on calling context to disable them accordingly. This PR refactors `ConfigureHDROutput` to use `CoreUtils.SetKeyword`.