Sg2/add shader samples (#7903)
* HDRP Nodes

* Added port tooltips for the eye nodes

* Updated AllShaderNodes and MigrationGroups shaders

* Updates based on Tracy's feedback

* removed the sampler port from the HDSampleBuffer node

* Weather Shaders

* Weather Textures

* Weather Shader Test Content

* Puddle Shader Update

* Project tracking shaders

* HDRP node tests

* Checking in my test content

* Saved a couple of files that I had open with my first submission

* Adding the Prototype documentation for our new subgraph nodes

Also removed the RGB port from the GatherTexture2D node

* Updates to NormalFromTexture and NormalUnpack

* Added a Multiplier input port to the Constant node

also updated some of the test content

* Made the Constant node's ports Vector instead of Float

* Added a UI Descriptor for the new Multiplier input port on the Constant node

* Test Preview Documentation

* Adding More Prototype Preview Documentation

also:  added ColorMode.uss, added SafeNormalize to the Normalize node.

* More new preview documentation files

* More Preview Documentation - and re-ordered parameter descriptors to match the port order

* More preview documentation

* More preview docs

* More preview documentation

* More preview documentation

* Updated and improved my docs so that lines start with a capital letter and end with a period.

* Noise3D Node and Noise Hash Node

* Renamed the Noise3D file

* Added Noise2D and textures, documentation for Noise 3D, improved matrix rotation

* Added documentation for Noise 2D and updated texture metadata

* Re-balanced the output of all noise types and added the sRGB checkbox. Also added the Scale control.

* Added Scale to Noise Hash. Added Noise Hash input defaults.  Added Noise Hash preview documentation

* Misc Updates

Improvements to Procedural Noise nodes
Changes the name of the output port on the ComputeWaterVertexPosition node to match the original version
Updated one of the colors in the ColorMode style sheet
Added a simplex texture
Re-added our new subgraph nodes in SG1 after the integration

* Updated Test Shader Files After Integration

Updated Test Shader Files After Integration
Updated display names of the REF data now that REF data dropdowns are working.
Added ParameterUIDescriptors for HDSceneColor and EmissionNode so that their static bool elements would appear.
Updated ObjectNode, and six water nodes that were changed in SG1 - so these updates were also needed in SG2.

* More updates to water and eye nodes

* Added OptionList dropdowns to nodes that were using default data

* Created the All Nodes graph

* Updated the matrix ReferenceValueDescriptors

This is done so that the Transformation Matrix node will work

* Added a ParameterUIDescriptor for the Type parameter on the Sample Texture 2D Nodes

* Added default values to The BlendWaterNormal node

* Updated EyeNodes and WaterNodes shader graph files

* Update to weather shaders

* Added the Fast version of ArcTangent2

* Added the Fast version to the ArcTangent2 node

* Adding missing md files

* Adding md file references to a few I missed.

* Renaming Nodes based on feedback from Anita and Martin

This set of nodes have been renamed to remove verbs from the title, remove abbreviations, and name the names more clear.

* Fixed a small typo in the WaterFoamData node

* Renamed one more node, renamed documentation files and documentation preview files to match the renamed nodes

* Updated synonyms so that the newly renamed nodes show up in the search for the old node names

* Added synonym for the Water Normal Blend node

* Removed In / Out Labels where they're not needed

Also fixed some port labels
Also removed Node Mode tags where nodes aren't using it.

* Added a missing label to the dropdown on the ArcTangent2 node

* Updated port names on Circle Pupil Animation Node ports based on Anita's feedback

* Updated preview documentation so node port names match the nodes

* Port Name Cleanup

A few more port name changes to fix a couple that I missed.

* Update com.unity.sg2/Editor/GraphDeltaRegistry/Registry/TransformNode.cs

Co-authored-by: Joe S. <joseph.savell@unity3d.com>

* Adding Node Example assets to ShaderGraphPlayground

These are all of the example assets that I've made so far.  They're in SG1 because the file format in SG2 is not yet stable. I also updated some of the texture assets so they have a smaller file size so that the example package can be smaller.

* Updates to node example assets

* Updated the dot product node

* Updates to node examples

* Checking in Latest Sample Assets

* Updating Blend and Parallax Mapping nodes with main function tags

The main function tags allow the nodes to specify an option that's not the first one as the default.

* Renamed And Reorganized Constant and Constructor Nodes

* Updated the documentation to go along with the Constant/Constructor changes that I made

* Delete ~UnityDirMonSyncFile~fb2cfc20dd054d849bf7ec7a09e013c9~

* Checking in my latest updates to the node examples

* Updates to sample content

* Reorganizing Example Content

I am creating a more organized folder structure for our example content library, so I've moved a lot of things around.

* Updates to Sample Content

Replaced the cobblestones textures with newer, better ones that I created in Substance Designer
Added the pattern examples from the existing sample content
Updated all files to be compatible with the latest version

* Hexagon Node

Adding the hexagon node to the node library - including docs
Added Noise Hash documentation
Fixed a typo in the Water Scattering Data preview
Renamed the Triplanar documentation to match the node's new name
Updated the documentation table of contents to match the new nodes and new node names we've been adding
Some small updates to the UV Flow Map documentation
Put the Light Texture 2D Node in the correct category
Added Offset to the UVFlowMap node in SG1
Updated the UVSphereMap node in SG1

* Delete test3.sg2

* Fixed a couple of typos in the hex grid code that Esme spotted

* Adds ice, stylized foliage, and sg urp lit shaders to ShaderGraphPlayGround

---------

Co-authored-by: bencloward <ben.cloward@unity3d.com>
Co-authored-by: Ben Cloward <88061215+bencloward@users.noreply.github.com>
Co-authored-by: Joe S. <joseph.savell@unity3d.com>