Add SRP code coverage pipeline - weekly
Currently there is no code coverage data available for SRP tests. This PR tries to introduce a pipeline that enables the code coverage functionality for the editmode and playmode SRP jobs that are QV candidates and is setup to run weekly to publish those results on the `coverage.hq.unity3d.com` and later on, on the codecov platform.