[RPW] Ensure Rendering Debugger runtime UI displays on top of other UI
Fix for https://jira.unity3d.com/browse/UUM-26782

Sort order for debug UI canvas was 0, so as soon as user has UI with larger sort order than that, it will be drawn on top of the debug UI that is intended to be an overlay on top of everything else.

To fix this, the sort order of DebugUICanvas.prefab is set to 32767 (INT16_MAX).
Similarly, the sort order of DebugUIPersistentCanvas.prefab is set to 32766 (INT16_MAX-1). This canvas is used when user "pins" debug values from debug UI with RShift, so they remain visible after the debug UI is closed. Having a value one smaller than the main UI ensures that the pinned values never occlude the main debug UI.

With the repro project
Before:
![image](https://media.github.cds.internal.unity3d.com/user/3380/files/39d8dd57-0857-4ef7-a10f-c8354eab7380)

After:
![image](https://media.github.cds.internal.unity3d.com/user/3380/files/4fce2470-38d0-407e-bead-829cdf2ca295)