[HDRP] Allow sorting of refractive objects with transparents
We want to be able to sort transparent objects that are in the refraction renderqueue with those in the pre refraction renderqueue (consider that water is a transparent with refraction, and clouds are transparents without refraction).
This is helpful when trying to render a transparent object that is crossing the water line for example.

A doc explaining the technical considerations is available here
https://docs.google.com/document/d/1g5P4CQq1wpgFjBDZzCBXb4hrAANqTtamv4oBKeAmM5w/edit

Transparents are now rendered in two separate color targets, normal and before refraction, and are later composited with refractive objects to be properly sorted. (This composition step is done at the same time as compositing low res transparents).

Additionally, refractive are rendered to a separate depth buffer to keep track of where they are.
(side effect is that gbuffer is rendered earlier in the frame, so it's now possible to sample the waterline node in custom pass from anything after and including the injection point `BeforePreRefraction`.

See result on screenshots below. It fixes looking through a window underwater, and looking at a transparent underwater from above water.
| trunk | PR|
| ------------- | ------------- |
| ![image](https://media.github.cds.internal.unity3d.com/user/2154/files/c93524f5-69cc-4a7a-af1a-381119f1ea8d) | ![image](https://media.github.cds.internal.unity3d.com/user/2154/files/e0c9df30-ad4c-467d-80be-facca9d985f9) | 
 | ![image](https://media.github.cds.internal.unity3d.com/user/2154/files/423ac3d6-0a44-438c-94c9-a3a5e1615a6b) | ![image](https://media.github.cds.internal.unity3d.com/user/2154/files/b1434e1e-529e-4ac9-a9ab-219cc69a32ed) |

Example of a transparent object that you can see though a refractive object (note that water is not visible cause it's a refractive through a refractive)
![image](https://media.github.cds.internal.unity3d.com/user/2154/files/f92dd23c-e1d8-4059-810a-0ac7f112e370)



Write doc about transparent prepass and caustics underwater