Transition HDRP/URP preview overlay to Cameras Overlay
With the new Cameras Overlay (#24971) in the SceneView, we aim at consolidating all existing camera preview overlays into one. As shown in the mentioned PR, we have added a new toolbar to the Camera Preview overlay to give the ability to artists to look through their camera, directly within the Scene View.

The `SceneViewCameraOverlay` class will be deprecated to leave place to this new overlay (see https://github.cds.internal.unity3d.com/unity/unity/pull/27013)

# HDRP
Currently, HDRP instantiates its own Overlay due to the fact that it has to handle the additional data. For the new Cameras Overlay, we are landing a change in the Camera Preview (#25452) to use the RenderRequest API so it can render a preview with any SRP that supports it.

This PR removes from HDRP the ability to generate its own Camera Preview. The new Cameras Overlay aims at supporting all RPs we support at Unity.

# URP

https://github.cds.internal.unity3d.com/unity/unity/pull/20864 has disconnected the URP CameraEditor from trunk's CameraEditor and had to implement a way to instantiate the SceneViewCameraOverlay with some C# reflection. This PR removes this as the new Cameras Overlay is the new replacement for this preview.