Platform/web/webgpu shader updates
Part of a set of PRs to merge the WebGPU development branch into trunk. This PR adds the shader changes for WebGPU, for both SRP and BiRP shaders.

Almost all of the changes is just adding SHADER_API_WEBGPU to shader includes.
The change to URP GaussianDepthOfField is needed because WebGPU has stricter validation than other APIs, and flagged it as a problem when the others did not catch it.