[HDRP]Stripping ComputeShaders when HDRP is disabled.
As there is not a built-in mechanism to tag a Compute Shader, in Cross Pipeline projects, or when building for Built-in when HDRP is installed. All the ComputeShaders from HDRP are not being stripped.

This makes the build time to increase by a lot, specially with the ~2k variants that the Deferred.compute is generating.

Making it unbearable  to build the project for URP when HDRP is installed.

Basically, this is a valid workarround until we have a way to know which ComputeShaders belong to HDRP without needing to fetch them from the Global Settings.