Fix issue with spurious SRP Changed callback invocation
Fix ISS bug https://jira.unity3d.com/browse/UUM-17548

The ISS bug is 21.3 but the issue was reproduced on trunk too.

The symptom of the bug is that all VFX graphs gets reimported when entering/leaving playmode but the cause is outside VFX area. It is due to a spurious call to SRP changed callback due to some algotithmic issue in the implementation.

Basically when a scene view is bypassing renderpipeline loop (because of the mode it's in or other reason), the active render pipeline is not set but the check is still performed which trigger a SRP changes callback invocation that triggers a reimport of VFX assets.