[RPW] Fix issue where renaming a VolumeProfile would break rendering
This PR fixes the following bug: https://jira.unity3d.com/browse/UUM-37604

The bug was introduced by this optimization PR: https://github.cds.internal.unity3d.com/unity/unity/pull/23193

When an asset is renamed in the editor, `OnEnable()` is called. The PR changed the code so that `parameterList` may not be empty when `OnEnable()` is called, resulting in `FindParameters()` simply appending more parameters to the list every time the asset is renamed. To fix the issue, we clear the list whenever `FindParameters()` is called.