Updated HDProbes to Support Custom Resolutions
The `Resolution` selection on Reflection Probes and Planar Reflection Probes where only respected when using the `Realtime` rendering mode. When using the `Custom` mode, the original size of the texture was used for atlasing and the only want to change that would be to updated the texture itself. Now each probe can dictate what resolution it should be used at for atlasing.