[SRP] Add Lens Flares properties to RP Assets to be able to strip shaders when not needed
The goal of this PR is to be able to strip lens flare shaders. ([slack discussion that started this](https://unity.slack.com/archives/C20E70CAC/p1670356509438579))
Before this PR, we'd have no way of knowing if the users needs those shaders so all the shaders variants were embed. 
With this PR, there's a property for each type of lens flare (Data Driven and Screen Space), enabled by default, to tell the stripper to pack the shaders or not. 

Here's in yellow what has been added to the HDRP Asset
![image](https://media.github.cds.internal.unity3d.com/user/1764/files/37f852ff-b168-4b59-ae1a-a7b26476526a)

Same on URP side (they don't use foldouts pattern so I didn't to respect coherence)
![image](https://media.github.cds.internal.unity3d.com/user/1764/files/b047662e-8929-47a6-a72b-67cd972b655e)

When the property is not checked, an error is shown on the SSLF override:
![image](https://media.github.cds.internal.unity3d.com/user/1764/files/0903a9f8-0d5a-4f55-af18-cd83eca1256f)

Nothing has been done on the SRP Lens Flare component since it's shared with URP. Waiting for @soufiane-khiat feedback on this one. 

Here's the log of the stripping now when both property are set to off. (on all assets)
![image](https://media.github.cds.internal.unity3d.com/user/1764/files/67010601-3f6a-478b-ad8b-d86e33607373)

Same URP Side
![image](https://media.github.cds.internal.unity3d.com/user/1764/files/9f233a71-c415-4f80-a74d-e83ce69f62c6)

As soon as any of the used RP Asset enables the properties all the shaders are not stripped anymore. 


- Here's a [video](https://unity.gyazo.com/4756ddff66415c628564f1017475119f) on switching between asset with none / one and both enabled (low / med / high)
- Here's [one](https://unity.gyazo.com/1323c4312c74b9c61e153d72fd1f1cad) with enabling / disabling both features on HDRP Asset
- Here's [another one](https://unity.gyazo.com/fd8ae570eccb34c6b187e54ec70198f6) URP side