changed some ScriptableRenderFeatures and UniversalRenderAsset API to…
Renderer and RendererData lists having only internal visibility has been a pain point for users in terms of URP expensibility and customization. Because of this limitation, the only workaround to this was to use reflection or having to fork the URP package.

This PR exposes safely those lists, allowing users to access those easily.

Some render features (SSAO and Decals) were also declared as internal classes by mistake, that is not consistent with all the other URP render features, so users wouldn't be able to check for those type ids, without having to use additional workarounds