[URP] Fixed Color Grading from Render Graph persisting outside of RG
On the Render Graph path, the internal color grading LUT is passed to the uber post processing through `Material.SetTexture`. On the non-Render Graph path, it's passed with `SetGlobalTexture`. The texture set by the RG takes precedence, even on the non-RG path, which caused the LUT used by post processing not to update when switching Render Graph on/off, along with switching HDR output on/off.

This PR modifies the non-RG path to use `Material.SetTexture` instead, so it correctly overrides the LUT set in the RG path.