[URP] Fixed Overlay UI rendering & debug views using the real camera size
Bug fix for [UUM-36231](https://jira.unity3d.com/browse/UUM-36231).
Screen space - Overlay UI, when rendered by the SRP, was using the camera intermediate target texture size rather than the real size. This meant the overlay UI was affected by the render scale, this is an issue when the scale mode is "Constant Pixel Size" or "Constant Physical Size". This happened for the HDR debug views too, creating multiple issues with cropped images.

To fix this, we pass the camera target's pixel size to the UI render target, instead of the camera target width/height that's been scaled. The handling of the depth render target had to be fixed as a consequence, we can't be using the intermediate depth target because the scale differs and it turns out the UI need a depth render target (it can read/write to the stencil for masking effects).
When rendering the overlay UI off-screen, we use a new `_OverlayUITexture_Depth` texture. When rendering on-screen, we use either the depth back buffer texture or the new `_DebugScreenDepth` texture (for the HDR debug views). Changes were made to the 2D renderer render graph path, so it keeps track of the depth back buffer for the UI on-screen pass.