[URP] Changed the PostPro passes injection points to more correct locations
The post processing passes are currently queued in the BeforeRenderingPostProcessing  injection point. This is wrong and doesn't allow users to queue their passes in between MotionVectorPass and PostProc.

More details in this forum post:

https://forum.unity.com/threads/urp-lacking-render-pass-injection-points.1440514/#post-9065386

This PR changes the injection points used by the PostPro passes to allow users to inject custom passes correctly before and after PostPro, if needed