2023.1: Fixed Rendering Debugger runtime UI getting occluded by user UI with sorting order greater than 0
The sort order of the debug UI canvas was initially set to 0. Therefore, whenever a user had a UI element with a higher sort order, it would be drawn on top of the debug UI, which was meant to overlay everything else.
This PR fixes the above issue.

Bug: UUM-26782
Backport: UUM-36608