Switch HDR SRGB encoding in SRPS to be display relative.
Switches how HDR SRGB values are encoded such that we move from treating 1.0 (the maximum we can output) as the reference white value which varies between platforms and is used for linear encoding to treating 1.0 as the maximum display output. SRGB encoding is generally only used with HDR when dealing with DisplayP3 where it should be max display brightness.