[HDRP]Fix recovering the current Quality level when migrating a HDRP Asset
Fix weird issue that the Quality Level is not the one from the previous version of Unity

This is due to when changing the SetQualityLevel the actual level on Quality Settings is being changed.

Fixing this issue: https://jira.unity3d.com/browse/UUM-37398

I've added a new API to better iterate over tiers and avoid this kind of problems.