Graphics/rpfoundation/nrp compiler2
This PR continues the work of adding native renderpass ("Vulkan Render Passes") support to the rendergraph. We previously expanded the existing rendergraph with new API to allow renderpasses to be expressed to the graph and updated URP to use this new API. This PR is the next step where the rendergraph will actually emit BeginRenderPass (& co) commands on the commandbuffer instead of SetRendertarget. When using the RenderPass API it will also emit appropriate load, store & resolve actions to ensure optimal performance on tile based hardware.

To achieve this a new render pass compiler backend is introduced that does correct usage tracking of resources. This new backend tracks in detail how render targets are used (as attachments or as random access textures, whether it needs resolved or individual MSAA sample data,...) and generates the optimal passes and load/store actions.