URP extraction testing
The purpose of this PR is to introduce intial tests which target meta passes without light baking. These tests show how to make more focused tests. As an example of this, we have had [a discussion on Slack](https://unity.slack.com/archives/C06TQ5M4J/p1686562874475879) in which we conclude that the less focused test `005_LitBakedEmission` caused false positives. The tests in this PR can be seen as a way to avoid spending time on such false positives in the future. Also, we are on a mission to substitute less focused tests, mainly graphics tests, with more focused tests. See for example https://jira.unity3d.com/browse/LIGHT-1366 - we will be able to land part of this when this PR lands.