[HDRP][APV]Fix leak generated by ComputeBuffer m_EmptyIndexBuffer
![image](https://media.github.cds.internal.unity3d.com/user/3279/files/6098e612-ab0e-453d-a382-358686ccea19)

The call to bind the APV runtime resources is not checking if the pipeline is using APV.

Making this call: `ProbeReferenceVolume.instance.BindAPVRuntimeResources(cmd, hdCamera.frameSettings.IsEnabled(FrameSettingsField.ProbeVolume))`

With : `hdCamera.frameSettings.IsEnabled(FrameSettingsField.ProbeVolume) == false`

Initializing `m_EmptyIndexBuffer`.

But the Cleanup of the APV system is only called when APV is enabled, therefore the `m_EmptyIndexBuffer `was not being disposed.