[Backport][2022.2][RPW]Stripping shaders from another SRP
Backporting : https://github.cds.internal.unity3d.com/unity/unity/pull/13647 and https://github.cds.internal.unity3d.com/unity/unity/pull/29985

Now we are including shaders in the following scenarios:

| Build Target Pipeline | Built-in | HDRP | URP |
|-----------------------|----------|------|-----|
| Built-in              |    :white_check_mark:      |    :x:  |  :x:   |
| HDRP                  |    :white_check_mark:      |    :white_check_mark:  |  :x:   |
| URP                   |     :white_check_mark:     |   :x:   |   :white_check_mark:  |

before:
| Build Target Pipeline | Built-in | HDRP | URP |
|-----------------------|----------|------|-----|
| Built-in              |    :white_check_mark:      |     :white_check_mark:   |   :white_check_mark:    |
| HDRP                  |    :white_check_mark:      |    :white_check_mark:  |   :white_check_mark:   |
| URP                   |     :white_check_mark:     |    :white_check_mark:   |   :white_check_mark:  |

As there is not a built-in mechanism to tag a Compute Shader, in Cross Pipeline projects, or when building for Built-in when HDRP is installed. All the ComputeShaders from HDRP are not being stripped.

This makes the build time to increase by a lot, specially with the ~2k variants that the Deferred.compute is generating.

Making it unbearable  to build the project for URP when HDRP is installed.

Basically, this is a valid workarround until we have a way to know which ComputeShaders belong to HDRP without needing to fetch them from the Global Settings.