Migrate metapass tests to an HDRP test project
The purpose of this PR is to introduce intial tests which target meta passes without light baking. These tests show how to make more focused and faster tests. The tests in this PR can be seen as a way to avoid spending time on false positives in the future. Also, we are on a mission to get rid of less focused and slow tests, mainly graphics tests.