[HDRP]Improving data layout and runtime performance for Custom PP Orders
As we are storing the Custom Post Process types as string ( AssemblyQualifiedName ) we are bound to do this:

```
var customPostProcessComponentType = Type.GetType(typeString);
                if (customPostProcessComponentType == null)
                    continue;
```

On every render.  profiling this call, we can see is not cheap:

![image](https://media.github.cds.internal.unity3d.com/user/3279/files/cb8359d2-fdfa-4a45-ad23-18bef0ce30bd)

Half of the cost of the parent method is spend on this call.

I've added an intermediate class that is the one that is being serialized, that serializes the strings to disk, but on runtime it is performed by type.

This avoids the call to `Type.GetType` for each Custom PP on each frame.

Also, introducing this custom class for storing the Custom PP, allows us to share all the UI code under a CustomPropertyDrawer. Making the UI code nicer and less coupled.