Migrate terrain metapass tests to URP test project
The purpose of this PR is to add tests which target meta passes for terrains without light baking. The tests in this PR can be seen as a way to avoid spending time on false positives in the future. Also, we are on a mission to get rid of less focused and slow tests, mainly graphics tests.