Revert the costly bugfix part of PR 28369
In https://github.cds.internal.unity3d.com/unity/unity/pull/28369 we replaced the value enum `FormatUsage` with a flags enum `GraphicsFormatUsage`. The majority of the PR is only refactoring code.

With this refactor-work we decided to fix some subtle bugs where mainly `FormatUsage.Linear | FormatUsage.Render` was being checked as if they were flags; in reality such a bitwise OR operation was actually checking `FormatUsage.Blend`. By now checking `GraphicsFormatUsage.Linear | GraphicsFormatUsage.Render` this bug was inherently tackled. However this resulted in a [performance hit](https://jira.unity3d.com/browse/UUM-39800) of several ms. It is assumed that this performance hit comes from the fact that the code "should have used 64bpp formats" but due to improper support checking it was unnecessarily falling back to lower-precision 32bpp formats. See also https://github.cds.internal.unity3d.com/unity/unity/pull/18904. The performance hit appears to be by design, but for users who don't observe any quality difference this could still resemble an unwanted performance "regression". This needs to be investigated to decide what the proper fix should be. Is the performance hit unavoidable? Should users now be more in control now that they are used to the performance of using 32bpp formats even if 64bpp was supported? Should the choice of going for 64bpp be reconsidered altogether given that the difference seems hardly noticeable? ...

The investigation of this issue is tracked under https://jira.unity3d.com/browse/UUM-41070, but for now it's preferred to stick to the original behavior until this is properly investigated. This PR reverts to that behavior, which today is more clearly wrong now that we have `GraphicsFormatUsage` as actual flags.