[shadergraph] improve graph concretization perf
https://jira.unity3d.com/browse/SGB-597 - Remove invalid characters from material subasset.
https://jira.unity3d.com/browse/SGB-613 - Improved perf issue w/modifications to a large graph.
https://jira.unity3d.com/browse/SGB-594 - Improved perf issue w/adding to large graph.
https://jira.unity3d.com/browse/SGB-605 - Fix for dropdowns not immediately propagating to open editor windows.
https://jira.unity3d.com/browse/SGB-592 - Fix for custom interpolator doc link.
https://jira.unity3d.com/browse/SGB-596 - Fix for Property Node doc link.
https://jira.unity3d.com/browse/SGB-531 - Fix CFN text field dimensions.
https://jira.unity3d.com/browse/SGB-581 - Fix CFN validation errors.
https://jira.unity3d.com/browse/SGB-561 - Fix for 'save' and 'save as' hotkey.


The main  purpose of the PR is to provide a hotfix to improve graph level concretization performance. Graph level concretization occurs on some user actions to finalize the topology of a node (notably, this is how dynamic vector ports are handled). When concretization for a node occurs, it's View receives a modified event and updates itself accordingly. It _also_ informs downstream node views that they may need to update as well. Because graph level concretization is already for all nodes, the concretization resulting in downstream nodes also getting notified was redundant in this case. I've added a flag to bypass this specific case until a better solution can be looked into.

The other issues are all low hanging fruit I bundled together for ease of porting and such.