[Backport][2022.3][VFX] Split Compute shader instancing batches when using different textures or graphic buffers
VFX Instancing allows to batch several VisualEffect components together to process their compute shaders with a single dispatch, and to render them together to save CPU time preparing the drawcalls.
Previous to this PR, it is possible to batch VisualEffect with different values in exposed properties, as long as those properties are not textures, meshes, or graphic buffers, that require different bindings. When a VisualEffectAsset detected a variable expression with type of one of those objects, instancing was disabled.
This PR enables instancing again for those cases, and allows the compute shaders to group instances based on the values of such variable objects in that specific compute shader. This PR does not handle this scenario for drawcalls in an optimal way, so they still have to be rendered one by one as if they were not batched.

![image](https://media.github.cds.internal.unity3d.com/user/2805/files/93053056-26ac-4aae-bb84-be73714a4021)


For instance, if there is a VisualEffectAsset with an exposed texture that is being used in the Update context, and a few VisualEffect of that type, with some of them overriding the texture, the batch will be processed like this:
- Init compute shader will be processed at the same time for all the instances, as it is not using the exposed texture.
- Update will be split in several dispatches, one for each different value of the exposed texture. Between each dispatch we only have to upload the instance active indirect and the instance prefix sum (if needed) and to bind the corresponding texture.
- OutputUpdate, if present, will also be processed at the same time for all the instances, if it is not using the exposed texture.
- Rendering will be batched for all instances, as the exposed texture in this example is not used in rendering.

![image](https://media.github.cds.internal.unity3d.com/user/2805/files/a2fc8d3a-3cb1-4a74-ba2d-0e5690a73f3b)
![image](https://media.github.cds.internal.unity3d.com/user/2805/files/f8d89471-49cb-41cf-a86e-d134225a9e9b)
![image](https://media.github.cds.internal.unity3d.com/user/2805/files/302d7986-b409-44af-a31f-5070a9047549)

However, if we have the same exposed texture being used by the Output context:
- Init and Update computes will be processed each of them as a single dispatch
- OutputUpdate will be split in as many dispatches as different texture values are used by the VisualEffects.
- Rendering of instances with the default value will be processed together. VisualEffect instances with an overriden value will be processed ONE BY ONE, regardless of how many of them are using the same texture. This will be addressed in a separate PR.

![image](https://media.github.cds.internal.unity3d.com/user/2805/files/0d345892-b26f-4e04-8ce9-c158321279f6)
![image](https://media.github.cds.internal.unity3d.com/user/2805/files/70053fc8-9592-464c-b52c-7f8c7ea05b54)
![image](https://media.github.cds.internal.unity3d.com/user/2805/files/ed536cd0-3e5a-4f40-91ef-77fd9fc8a669)

Apart from those changes, this PR includes some improvements to VFX culling and some fixes.